separate value numbering
